Separate builder bid into a strategy.
