+++ services/blockrelay/standard/service.go
@@ -30,10 +30,10 @@ import (
 	"github.com/attestantio/vouch/services/chaintime"
 	"github.com/attestantio/vouch/services/metrics"
 	"github.com/attestantio/vouch/services/signer"
+	"github.com/attestantio/vouch/strategies/builderbid"
 	"github.com/pkg/errors"
 	"github.com/rs/zerolog"
 	zerologger "github.com/rs/zerolog/log"
-	e2types "github.com/wealdtech/go-eth2-types/v2"
 	"github.com/wealdtech/go-majordomo"
 	"golang.org/x/sync/semaphore"
 )
@@ -59,15 +59,13 @@ type Service struct {
 	signedValidatorRegistrationsMu            sync.RWMutex
 	secondaryValidatorRegistrationsSubmitters []consensusclient.ValidatorRegistrationsSubmitter
 	logResults                                bool
-	applicationBuilderDomain                  phase0.Domain
 	releaseVersion                            string
+	builderBidProvider                        builderbid.Provider
 
 	executionConfig   blockrelay.ExecutionConfigurator
 	executionConfigMu sync.RWMutex
 
-	relayPubkeys   map[phase0.BLSPubKey]*e2types.BLSPublicKey
-	relayPubkeysMu sync.RWMutex
-	activitySem    *semaphore.Weighted
+	activitySem *semaphore.Weighted
 }
 
 // module-wide log.
@@ -90,24 +88,6 @@ func New(ctx context.Context, params ...Parameter) (*Service, error) {
 		return nil, errors.New("failed to register metrics")
 	}
 
-	// The application domain is static, so fetch it here once.
-	spec, err := parameters.specProvider.Spec(ctx)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to obtain spec")
-	}
-	tmp, exists := spec["DOMAIN_APPLICATION_BUILDER"]
-	if !exists {
-		return nil, errors.New("failed to obtain application builder domain type")
-	}
-	applicationBuilderDomainType, ok := tmp.(phase0.DomainType)
-	if !ok {
-		return nil, errors.New("unexpected type for application builder domain type")
-	}
-	domain, err := parameters.domainProvider.GenesisDomain(ctx, applicationBuilderDomainType)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to obtain application builder domain")
-	}
-
 	s := &Service{
 		monitor:                      parameters.monitor,
 		majordomo:                    parameters.majordomo,
@@ -124,13 +104,12 @@ func New(ctx context.Context, params ...Parameter) (*Service, error) {
 		timeout:                      parameters.timeout,
 		signedValidatorRegistrations: make(map[phase0.Root]*apiv1.SignedValidatorRegistration),
 		secondaryValidatorRegistrationsSubmitters: parameters.secondaryValidatorRegistrationsSubmitters,
-		logResults:               parameters.logResults,
-		applicationBuilderDomain: domain,
-		releaseVersion:           parameters.releaseVersion,
-		builderBidsCache:         make(map[string]map[string]*builderspec.VersionedSignedBuilderBid),
-		relayPubkeys:             make(map[phase0.BLSPubKey]*e2types.BLSPublicKey),
-		executionConfig:          &v2.ExecutionConfig{Version: 2},
-		activitySem:              semaphore.NewWeighted(1),
+		logResults:         parameters.logResults,
+		releaseVersion:     parameters.releaseVersion,
+		builderBidsCache:   make(map[string]map[string]*builderspec.VersionedSignedBuilderBid),
+		executionConfig:    &v2.ExecutionConfig{Version: 2},
+		activitySem:        semaphore.NewWeighted(1),
+		builderBidProvider: parameters.builderBidProvider,
 	}
 
 	// Carry out initial fetch of execution configuration.
