+++ strategies/builderbid/best/parameters.go
@@ -1,0 +1,128 @@
+// Copyright © 2023 Attestant Limited.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// package deadline is a strategy that obtains builder bids from multiple
+// relays repeatedly up to the supplied deadline into a slot.
+package best
+
+import (
+	"time"
+
+	consensusclient "github.com/attestantio/go-eth2-client"
+	"github.com/attestantio/vouch/services/chaintime"
+	"github.com/attestantio/vouch/services/metrics"
+	nullmetrics "github.com/attestantio/vouch/services/metrics/null"
+	"github.com/pkg/errors"
+	"github.com/rs/zerolog"
+)
+
+type parameters struct {
+	logLevel       zerolog.Level
+	monitor        metrics.Service
+	specProvider   consensusclient.SpecProvider
+	domainProvider consensusclient.DomainProvider
+	chainTime      chaintime.Service
+	timeout        time.Duration
+	releaseVersion string
+}
+
+// Parameter is the interface for service parameters.
+type Parameter interface {
+	apply(*parameters)
+}
+
+type parameterFunc func(*parameters)
+
+func (f parameterFunc) apply(p *parameters) {
+	f(p)
+}
+
+// WithLogLevel sets the log level for the module.
+func WithLogLevel(logLevel zerolog.Level) Parameter {
+	return parameterFunc(func(p *parameters) {
+		p.logLevel = logLevel
+	})
+}
+
+// WithMonitor sets the monitor for the service.
+func WithMonitor(monitor metrics.Service) Parameter {
+	return parameterFunc(func(p *parameters) {
+		p.monitor = monitor
+	})
+}
+
+// WithSpecProvider sets the spec provider.
+func WithSpecProvider(provider consensusclient.SpecProvider) Parameter {
+	return parameterFunc(func(p *parameters) {
+		p.specProvider = provider
+	})
+}
+
+// WithDomainProvider sets the signature domain provider.
+func WithDomainProvider(provider consensusclient.DomainProvider) Parameter {
+	return parameterFunc(func(p *parameters) {
+		p.domainProvider = provider
+	})
+}
+
+// WithChainTime sets the chaintime service.
+func WithChainTime(service chaintime.Service) Parameter {
+	return parameterFunc(func(p *parameters) {
+		p.chainTime = service
+	})
+}
+
+// WithTimeout sets the timeout for requests.
+func WithTimeout(timeout time.Duration) Parameter {
+	return parameterFunc(func(p *parameters) {
+		p.timeout = timeout
+	})
+}
+
+// WithReleaseVersion sets the release version for Vouch.
+func WithReleaseVersion(version string) Parameter {
+	return parameterFunc(func(p *parameters) {
+		p.releaseVersion = version
+	})
+}
+
+// parseAndCheckParameters parses and checks parameters to ensure that mandatory parameters are present and correct.
+func parseAndCheckParameters(params ...Parameter) (*parameters, error) {
+	parameters := parameters{
+		logLevel: zerolog.GlobalLevel(),
+		monitor:  &nullmetrics.Service{},
+	}
+	for _, p := range params {
+		if params != nil {
+			p.apply(&parameters)
+		}
+	}
+
+	if parameters.monitor == nil {
+		return nil, errors.New("no monitor specified")
+	}
+	if parameters.specProvider == nil {
+		return nil, errors.New("no spec provider specified")
+	}
+	if parameters.domainProvider == nil {
+		return nil, errors.New("no domain provider specified")
+	}
+	if parameters.chainTime == nil {
+		return nil, errors.New("no chain time specified")
+	}
+	if parameters.timeout == 0 {
+		return nil, errors.New("no timeout specified")
+	}
+
+	return &parameters, nil
+}
