+++ strategies/builderbid/service.go
@@ -1,0 +1,36 @@
+// Copyright © 2023 Attestant Limited.
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package builderbid
+
+import (
+	"context"
+
+	"github.com/attestantio/go-block-relay/services/blockauctioneer"
+	"github.com/attestantio/go-eth2-client/spec/phase0"
+	"github.com/attestantio/vouch/services/beaconblockproposer"
+)
+
+// Provider is the interface for providing builder bids.
+type Provider interface {
+	// BuilderBid returns a builder bid.
+	BuilderBid(ctx context.Context,
+		slot phase0.Slot,
+		parentHash phase0.Hash32,
+		pubkey phase0.BLSPubKey,
+		proposerConfig *beaconblockproposer.ProposerConfig,
+	) (
+		*blockauctioneer.Results,
+		error,
+	)
+}
